Agentic AI GIKI Final/requirements.txt
agno>=1.7.0
groq>=0.8.0
duckduckgo-search>=6.1.3
chromadb>=0.5.4
sentence-transformers>=3.0.1
pypdf>=4.2.0
python-dotenv>=1.0.1
datasets>=3.0.0
tqdm>=4.66.4
tabulate>=0.9.0
numexpr>=2.10.1
datasets>=2.17.0
huggingface_hub>=0.23.0
pandas>=2.0.0
pyarrow>=14.0.0
